requirements.txt
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==2.6.2
Pillow==6.0.0
pytesseract==0.2.6
xlwt==1.3.0
